Implement user authorization logic on routes
Connect to backend to ensure user validation and authentication.

/face-detection route won't be accessible unless the user is
authenticated and authorized in the backend.

Routes: /, /register, /error won't be accessible once the user is
authenticated.

User can logout when clicking on the logout button in the navbar. This
only removes the user token cookie to remove authorization from the
backend.

Create utilities folder to contain React Routers actions and loaders.