Update face-detection page
Change the detection input from text (URL) to file. This is intended to
allow users to upload a picture containing the celebrity they want to
recognize.

Make a custom input type=file according to the page style.

Display the image preview once user selects a file.

IMPORTANT: Initial connection to back end allowing upload of the image
to the server. When a positive response is recieved, the page should
take the data in order to display the celebrity face bounding box and
name, then refresh the user entries.