Fix onFileChange bug on faceDetection page
The bug was trigger by the user when clickend on the file input and
then click on cancel or select no image. This would prevent the user
from submit the form but the previous image was still shown in the UI.

The fix consist in hiding the image when the file input is empty.